+++ packages/colony-js-utils/src/isValidAddress.js
@@ -1,7 +1,7 @@
 /* @flow */
 
 import { isAddress } from 'web3-utils';
-import { NON_EXISTENT_ADDRESS } from './constants';
+import isEmptyHexString from './isEmptyHexString';
 
 /**
  * Given an input, return true if it is a valid address.
@@ -13,6 +13,6 @@ export default function isValidAddress(address: any): boolean {
   return (
     typeof address === 'string' &&
     isAddress(address) &&
-    address !== NON_EXISTENT_ADDRESS
+    !isEmptyHexString(address)
   );
 }